added game over stage

+++ GameServer/startServer.sh
@@ -11,12 +11,22 @@ read -e -p "Enter participant id: " -i "" PARTICIPANTID
 read -e -p "Enter robot address: " -i "" ROBOTADDRESS
 #echo $ROBOTADDRESS
 
+if [[ -z "$CONDITION" ]]; then
+   echo "you did not name the CONDITION"
+   exit
+fi
+
+if [[ -z "$PARTICIPANTID" ]]; then
+   echo "you did not name the PARTICIPANTID"
+   exit
+fi
+
 #echo "iojs --debug --harmony_classes server/index.js --condition $CONDITION --pId $PARTICIPANTID --robotAddress $ROBOTADDRESS"
 
 #read -e -p "Press any Key to start"
 
 trap "kill 0" SIGINT
-(iojs --debug --harmony_classes server/index.js --condition $CONDITION --pId $PARTICIPANTID --robotAddress $ROBOTADDRESS) & (grunt serve --force)
+(iojs --debug --harmony_classes server/index.js --condition $CONDITION --pId $PARTICIPANTID --robotAddress $ROBOTADDRESS) & (grunt serve --force) & (./helperScripts/startAudioRecroding.sh $CONDITION $PARTICIPANTID)
 #(python naoServer/index.py) & (grunt nodemon) & (grunt serve)
 
 read -e -p "Press Ctrl + C to end"
